requirements.txt
flask
flask-discord
py-cord
python-dotenv
gunicorn
pymongo
flask-socketio
redis
python-socketio[client]
eventlet